+++ totp/workers/type-code-common.c
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "type-code-common.h"
 #include <furi/furi.h>
 #include <furi_hal.h>
 #include "../../services/convert/convert.h"
