+++ totp/workers/type-code-common.h
