+++ totp/workers/usb_type_code/usb_type_code.c
@@ -1,7 +1,7 @@
 #include "usb_type_code.h"
 #include "../../services/convert/convert.h"
 #include "../../types/token_info.h"
-#include "../common.h"
+#include "../type-code-common.h"
 
 static void totp_type_code_worker_restore_usb_mode(TotpUsbTypeCodeWorkerContext* context) {
     if(context->usb_mode_prev != NULL) {
